+++ models/br_inep_sinopse_estatistica_educacao_basica/br_inep_sinopse_estatistica_educacao_basica__educacao_especial_localizacao.sql
@@ -18,7 +18,7 @@ select
     safe_cast(tipo_classe as string) tipo_classe,
     safe_cast(rede as string) rede,
     safe_cast(localizacao as string) localizacao,
-    safe_cast(quantidade_matricula as int64) quantidade_matricula,
+    safe_cast(quantidade_matricula as numeric) quantidade_matricula,
 from
     `basedosdados-staging.br_inep_sinopse_estatistica_educacao_basica_staging.educacao_especial_localizacao`
     as t
